% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap_france.R
\name{basemap_france}
\alias{basemap_france}
\title{Create a basemap of metropolitan France}
\usage{
basemap_france(fill = "#F0F0F0", color = "white", corse = FALSE)
}
\arguments{
\item{fill}{Fill color}

\item{color}{Borders color}

\item{corse}{If TRUE, Corse is included into the basemap (default: FALSE)}
}
\value{
a ggplot layer representing the metropolitan France (WGS84 coords)
}
\description{
Create ggplot layer representing the metropolitan France.
}
\examples{
library(ggplot2)
data <- data.frame(x = c(2.3, -2.5), y = c(48.8, 48.5), couleur = c("red","green"))
basemap_france() +
  geom_point(data = data, aes(x = x, y = y, color = couleur))

}
