% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorise_log.R
\name{categorize_log}
\alias{categorize_log}
\title{Logarithmic categorization of a numeric variable}
\usage{
categorize_log(
  df,
  column_to_categorize,
  number_of_intervals = 10,
  new_column_name = "categories"
)
}
\arguments{
\item{df}{Dataframe to complete}

\item{column_to_categorize}{A numeric column of the dataframe, containing only positive numbers}

\item{number_of_intervals}{Number of categories to create}

\item{new_column_name}{Name of the new column (default : categories)}
}
\value{
A dataframe equals to the input, with one more column
}
\description{
Add a column to a dataframe, defined as a logarithmic categorization of a numeric variable you chose in this dataframe.
}
\examples{
library(dplyr)
data <- tibble(observations = 1:100, valeurs = rnorm(100, mean = 100, sd = 50))
data <- data \%>\%
categorize_log(column_to_categorize = observations,
new_column_name = categories,
number_of_intervals = 5)


}
