% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_vector.R
\name{color_vector}
\alias{color_vector}
\title{Color vector compliant with the graphical charter of the ART}
\usage{
color_vector(...)
}
\arguments{
\item{...}{Zero, one or more character strings representing colors of the graphical charter or other colors in hexadecimal format}
}
\value{
A vector containing \strong{color hexadecimal values}. Without argument, the function returns a vector composed of each of the colors of the graphic charter. With arguments, the function returns a vector with all the hexadecimal code of the arguments.
}
\description{
Create a color vector containing the hexadecimal values of the ART graphical charter
}
\examples{
color_vector()
color_vector("blue", "#00000")
color_vector("#00000")

}
