% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_from_pgsql.R
\name{import_data_from_pgsql}
\alias{import_data_from_pgsql}
\title{Import data from PostgreSQL}
\usage{
import_data_from_pgsql(
  sql,
  host,
  port,
  dbname,
  env_var_usr = "database_usr",
  env_var_pwd = "database_pwd"
)
}
\arguments{
\item{sql}{A SQL character string requesting the database}

\item{host}{A character string naming the host}

\item{port}{A character string setting the port number}

\item{dbname}{A character string naming the database}

\item{env_var_usr}{Name of the environment variable storing your database username}

\item{env_var_pwd}{Name of the environment variable storing your database password}
}
\value{
a dataframe
}
\description{
Import a data table from a PostgreSQL database.

By creating two environment variables called \strong{database_usr} storing your database username and \strong{database_pwd} storing your database password, you will get much faster !
}
\examples{
data <- import_data_from_pgsql("SELECT * FROM information_schema.schemata")

}
