% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_meta_from_pgsql.R
\name{import_meta_from_pgsql}
\alias{import_meta_from_pgsql}
\title{Import metadata from PostgreSQL}
\usage{
import_meta_from_pgsql(
  schema,
  table,
  host,
  port,
  dbname,
  env_var_usr = "database_usr",
  env_var_pwd = "database_pwd"
)
}
\arguments{
\item{schema}{a character string naming the schema}

\item{table}{a character string naming the table}

\item{host}{a character string naming the host}

\item{port}{a character string setting the port number}

\item{dbname}{a character string naming the database}

\item{env_var_usr}{Name of the environment variable storing your database username}

\item{env_var_pwd}{Name of the environment variable storing your database password}
}
\value{
a dataframe
}
\description{
Import columns description of a table stored in a PostgreSQL database.

By creating a variable environment \strong{database_usr} for your database username and \strong{database_pwd} for your database password, you will improve your experience !
}
\examples{
meta <- import_meta_from_pgsql("fer_circulation", "circulation_quotidienne")

}
