% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_data_into_pgsql.R
\name{insert_data_into_pgsql}
\alias{insert_data_into_pgsql}
\title{Insert data into PostgreSQL}
\usage{
insert_data_into_pgsql(
  df,
  schema,
  table,
  delete_previous_records = FALSE,
  host,
  port,
  dbname,
  env_var_usr = "database_usr",
  env_var_pwd = "database_pwd"
)
}
\arguments{
\item{df}{a dataframe to be inserted into a PostgreSQL table}

\item{schema}{a character string naming the PostgreSQL schema}

\item{table}{a character string naming the PostgreSQL table}

\item{delete_previous_records}{TRUE if you want to delete previous records of the table}

\item{host}{a character string naming the host}

\item{port}{a character string setting the port number}

\item{dbname}{a character string naming the database}

\item{env_var_usr}{Name of the environment variable storing your database username}

\item{env_var_pwd}{Name of the environment variable storing your database password}
}
\value{
TRUE if the data is correctly inserted
}
\description{
Insert data into an existing table in a PostgreSQL database.

By creating two environment variables called \strong{database_usr} storing your database username and \strong{database_pwd} storing your database password, you will get much faster !
}
\examples{
library(dplyr)
mtcars \%>\% insert_data_into_pgsql(schema = "public", table = "mtcars", delete_previous_records = TRUE)

}
