% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isochrones.R
\name{isochrones}
\alias{isochrones}
\title{Compute isochrones around a set of points}
\usage{
isochrones(data_sf, breaks = c(0, 15, 30, 60), res = 30)
}
\arguments{
\item{data_sf}{a SF dataframe containing a set of points}

\item{breaks}{a vector of numeric values corresponding to the drive times of the isochrones (default: c(0,15,30,60)).}

\item{res}{number of points used to compute isochrones, one side of the square grid, the total number of points will be res*res (default: 30).}
}
\value{
a SF object containing one geometry for each break (the isochrones of all of the points are joined).
}
\description{
Compute isochrones around a set of points listed in a SF dataframe.
}
\examples{
library(dplyr)
library(sf)
library(ggplot2)
data <- tibble(x = c(2.3, -2.5), y = c(48.8, 48.5)) \%>\%
  st_as_sf(coords = c("x", "y"))
isochrones <- data \%>\% isochrones(breaks = c(10,30,90))
basemap_france() +
  geom_sf(data = isochrones, aes(fill = max))

}
