% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{pointplot}
\alias{pointplot}
\title{Point plot compliant with corporate identity}
\usage{
pointplot(
  data,
  x,
  y,
  color,
  shape,
  size,
  size.fixed = 1.5,
  facet,
  color.values = NULL,
  x.accuracy = 1,
  x.suffix = NULL,
  y.accuracy = 1,
  y.suffix = NULL,
  label,
  label.filter,
  label.size = 3,
  label.color = NULL,
  label.hjust = 0,
  label.vjust = 0,
  label.repel.force = 1,
  title = NULL,
  subtitle = NULL,
  x.name = NULL,
  y.name = NULL,
  legend.color.name = NULL,
  legend.shape.name = NULL,
  legend.size.name = NULL
)
}
\arguments{
\item{data}{A dataframe}

\item{x}{A numeric column of the dataframe used for x axis.}

\item{y}{A numeric column of the dataframe used for y axis.}

\item{color}{(optional) A categorical column of the dataframe used to create a color legend.}

\item{shape}{(optional) A categorical column of the dataframe used to create a shape legend.}

\item{size}{(optional) A numeric column of the dataframe used to create a size legend.}

\item{size.fixed}{Size of the points. Used only if size is missing (default: 1.5).}

\item{facet}{(optional) A categorical column of the dataframe. One chart will be created for each category.}

\item{color.values}{A vector of colors. If named, then the values will be matched based on the names.}

\item{x.accuracy}{A number to round the x axis text to. Use (e.g.) 0.01 to show 2 decimal places of precision. If NULL, uses a heuristic that should ensure breaks have the minimum number of digits needed to show the difference between adjacent values (default: 1).}

\item{x.suffix}{A character string used as a suffix for x axis text.}

\item{y.accuracy}{A number to round the y axis text to. Use (e.g.) 0.01 to show 2 decimal places of precision. If NULL, uses a heuristic that should ensure breaks have the minimum number of digits needed to show the difference between adjacent values (default: 1).}

\item{y.suffix}{A character string used as a suffix for y axis text.}

\item{label}{A categorical column of the dataframe used to label the points.}

\item{label.filter}{A character string used to filter the points to be labelled (e.g. if label.filter = "column_name == 2", then only observations where column_name is equal to 2 will be labelled).}

\item{label.size}{A numeric value indicating the labels size.}

\item{label.color}{A character string used to set the label color. By default, labels are the same color as points if \strong{color} is filled, or black if \strong{color} is missing.}

\item{label.hjust}{A numeric value used for the horizontal adjustment of the label position. The unit is the same as for the data unit on the x-axis.}

\item{label.vjust}{A numeric value used for the vertical adjustment of the label position. The unit is the same as for the data unit on the y-axis.}

\item{label.repel.force}{A numeric value used as force of repulsion between overlapping text labels (defaults to 1).}

\item{title}{A character string for the title of the chart.}

\item{subtitle}{A character string for the subtitle of the chart.}

\item{x.name}{A character string for the x axis title.}

\item{y.name}{A character string for the y axis title.}

\item{legend.color.name}{A character string for the color legend title.}

\item{legend.shape.name}{A character string for the shape legend title.}

\item{legend.size.name}{A character string for the size legend title.}
}
\value{
A ggplot2 pointplot (geom_point) compliant with the corporate identity.
}
\description{
A point plot compliant with the corporate identity.
}
\examples{
library(dplyr)
mtcars \%>\%
  pointplot(x = mpg,
                y = qsec,
                x.name = "Miles per gallon",
                y.name = "1/4 mile time",
                title = "Mile time",
                subtitle = "According to miles per gallon")
#
mtcars \%>\%
  mutate(am = case_when(am == 1 ~ "Automatic", TRUE ~ "Manual")) \%>\%
  pointplot(x = mpg,
                y = qsec,
                color = cyl,
                label = gear,
                label.filter = "cyl == 6",
                x.name = "Miles per gallon",
                y.name = "1/4 mile time",
                legend.color.name = "Cylinders",
                title = "Mile time",
                subtitle = "According to miles per gallon")

}
