% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_gradient.R
\name{scale_gradient}
\alias{scale_gradient}
\alias{scale_color_gradient}
\alias{scale_fill_gradient}
\title{Gradient color scales}
\usage{
scale_color_gradient(
  low = "lightgreen",
  mid = "orange",
  high = "lightred",
  ...
)

scale_fill_gradient(low = "lightgreen", mid = "orange", high = "lightred", ...)
}
\arguments{
\item{low}{Color for low end of the gradient (default: lightgreen)}

\item{mid}{Color for the middle of the gradient (default: orange)}

\item{high}{Color for high end of the gradient (default: lightred)}

\item{...}{Any other argument accepted by \strong{scale_colour_gradient} or \strong{scale_fill_gradient} from ggplot2}
}
\value{
A ggplot2 gradient scale, just like \strong{scale_color_gradient} or \strong{scale_fill_gradient}
}
\description{
Create a 3 color gradient (low-mid-high) compliant with the corporate identity.
}
\examples{
library(ggplot2)
library(dplyr)
mtcars \%>\%
  ggplot() +
  aes(x = mpg, y = disp, color = mpg) +
  scale_color_gradient() +
  geom_point() +
  theme()

}
