% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_manual.R
\name{scale_manual}
\alias{scale_manual}
\alias{scale_color_manual}
\alias{scale_fill_manual}
\title{Create your own discrete scale}
\usage{
scale_color_manual(values = unname(datamanir::color_vector()), ...)

scale_fill_manual(values = unname(datamanir::color_vector()), ...)
}
\arguments{
\item{values}{A vector of character strings corresponding to corporate colors or other hexadecimal values (default: a set of corporate colors)}

\item{...}{Any other argument accepted by \strong{scale_color_manual} ou \strong{scale_fill_manual} from ggplot2}
}
\value{
A discrete scale, just like \strong{scale_color_manual} ou \strong{scale_fill_manual} from ggplot2
}
\description{
These functions allow you to specify your own set of mappings from levels in the data to aesthetic values, using corporate colors.
}
\examples{
library(dplyr)
library(ggplot2)
#
p <- ggplot(mtcars, aes(mpg, wt)) +
geom_point(aes(colour = factor(cyl)))
#
colors <- c("red","blue","#762147")
names(colors) <- c("6", "4", "8")
p + scale_color_manual()
#
p + scale_color_manual(values = colors, labels = c("four", "six", "heigth"), name = "Cylinders")

}
