% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme}
\alias{theme}
\title{Corporate theme}
\usage{
theme(
  text = element_text(colour = "#1d1d1b", family = "Century Gothic", size = 10),
  plot.title = element_text(colour = "#1d1d1b", size = 11),
  plot.subtitle = element_text(colour = "#1d1d1b", size = 9, vjust = 4),
  panel.border = element_blank(),
  panel.background = element_blank(),
  panel.grid.major.x = element_line(colour = "#E1DDDD", size = 0.15),
  panel.grid.minor.x = element_blank(),
  panel.grid.major.y = element_line(colour = "#E1DDDD", size = 0.15),
  panel.grid.minor.y = element_blank(),
  axis.title.x = element_blank(),
  axis.title.y = element_blank(),
  legend.title = element_text(colour = "#1d1d1b", size = 10),
  legend.position = "right",
  rect = element_blank(),
  axis.ticks = element_blank(),
  ...
)
}
\arguments{
\item{...}{See \strong{theme} from \strong{ggplot2}}
}
\value{
A ggplot2 theme with default values compliant with the corporate identity.
}
\description{
A theme compliant with the corporate identity. At first use, it may be necessary to import Windows font styles (see examples).
}
\examples{
library(ggplot2)
ggplot() + aes(x = c(1,2,3), y = c(1,2,3)) + geom_point() + theme_art()
# At first use, it may be necessary to import Windows font styles:
library(extrafont)
font_import()

}
